backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Anton Melser <anton@melser.org>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
fastapi = "^0.68.0"
python-multipart = "^0.0.5"
email-validator = "^1.1.3"
requests = "^2.26.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.0.1"
pydantic = "^1.8.2"
emails = "^0.6"
raven = "^6.10.0"
gunicorn = "^20.1.0"
Jinja2 = "^3.0.1"
psycopg2-binary = "^2.9.1"
alembic = "^1.6.5"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.23"}
python-jose = "^3.3.0"
asyncpg = "^0.24.0"
httpx = "^0.19.0"
strawberry-graphql = "^0.73.2"
aiohttp-retry = "^2.4.5"
aiohttp = "^3.7.4"
broadcaster = {extras = ["postgres"], version = "^0.2.0"}
webvtt-py = "^0.4.6"
dawn = "^0.11.0"
bs4 = "^0.0.1"
aiokafka = "^0.7.1"
pymemcache = "^3.5.0"
mmh3 = "^3.0.0"
OpenCC = "1.1.1.post1"
python-magic = "^0.4.24"
orjson = "^3.6.3"
logging_tree = "^1.9"
aiofiles = "^0.7.0"
pytz = "^2021.1"
faust-streaming = { git = "https://github.com/AntonOfTheWoods/faust.git", branch = "debogue", extras = ["rocksdb"] }

[tool.poetry.dev-dependencies]
mypy = "^0.910"
autoflake = "^1.4"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-asyncio = "^0.15.1"
pre-commit = "^2.13.0"
pylint = "^2.10.2"
black = "^21.7b0"

[tool.pylint.master]
extension-pkg-whitelist = ["pydantic"]

[tool.pylint.messages_control]
# I0011 Warning locally suppressed using disable-msg
# disabling wrong-import-order because isort is run in pre-commit, pylint ignores [tool.isort] in this file
# and I can't work out how to get pylint to consider 'known-first-party'
disable = """
    invalid-name,
    logging-fstring-interpolation,
    fixme,
    I0011,
    missing-docstring,
    C0330,
    too-few-public-methods,
    too-many-arguments,
    C0326,
    pointless-string-statement,
    no-member,
    R0901,
    R0801,
    wrong-import-order,
    too-many-instance-attributes,
"""

[tool.pylint.basic]
# Good variable names which should always be accepted, separated by a comma.
good-names = "i,j,k,ex,_,f,e,s,w,li,t"

[tool.pylint.format]
max-line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
